Dividing by complex pilots in UE channel estimation.
Former-commit-id: 587b8709a771b391b6ac69115957ea6144d684d5

+++ src/client/phy-ue.hpp
@@ -369,6 +369,6 @@ private:
     int max_equaled_frame = 0;
     // long long* demul_output;
     // float* equal_output;
-    int record_frame = -1;
+    size_t record_frame = -1;
 };
 #endif
